requirements.txt
aiogram==2.25.1
openai==1.13.3
easyocr
python-docx
pymupdf
python-dotenv